requirements.txt
jsonpath==0.82
requests==2.28.0
PyHamcrest==2.0.3
pytest~=7.1.2